refactor: DTO의 위치를 알맞게 이동
