+++ src/main/java/wad/seoul_nolgoat/domain/bookmark/BookmarkRepositoryCustom.java
@@ -2,7 +2,7 @@ package wad.seoul_nolgoat.domain.bookmark;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import wad.seoul_nolgoat.web.store.dto.response.StoreForBookmarkDto;
+import wad.seoul_nolgoat.web.bookmark.dto.response.StoreForBookmarkDto;
 
 public interface BookmarkRepositoryCustom {
