+++ src/main/java/wad/seoul_nolgoat/service/bookmark/BookmarkService.java
@@ -12,7 +12,7 @@ import wad.seoul_nolgoat.domain.store.StoreRepository;
 import wad.seoul_nolgoat.domain.user.User;
 import wad.seoul_nolgoat.domain.user.UserRepository;
 import wad.seoul_nolgoat.exception.ApiException;
-import wad.seoul_nolgoat.web.store.dto.response.StoreForBookmarkDto;
+import wad.seoul_nolgoat.web.bookmark.dto.response.StoreForBookmarkDto;
 
 import static wad.seoul_nolgoat.exception.ErrorCode.*;
